refactor: rename filters to exportInitProps
